Refactor: refactoring Attributes
